pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spark-perf-lint"
version = "0.1.0"
description = "Enterprise-grade Apache Spark performance linter"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Sucandra Dasa" }]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.25",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.9",
    "pre-commit>=3.0",
]

[project.scripts]
spark-perf-lint = "spark_perf_lint.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: slow performance tests (may take several seconds)",
    "tier2: requires a live LLM API key (Anthropic)",
    "tier3: requires a running Spark runtime",
]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
